+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
+- [#10](https://github.com/zendframework/zend-mvc-plugin-flashmessenger/pull/10) ports the flashMessenger view helper from zend-view to this package, under the
+  class name `Zend\Mvc\Plugin\FlashMessenger\View\Helper\FlashMessenger`, and using the helper name
+  `flashMessenger()`.
+
 - [#19](https://github.com/zendframework/zend-mvc-plugin-flashmessenger/pull/19) adds support for PHP 7.1 and 7.2.
 
 ### Changed